DTO and processing logic

+++ Solution/Shared/KeyEventMessage.cs
@@ -6,7 +6,7 @@ namespace Shared
     public class KeyEventMessage
     {
         [Key(0)]
-        public int KeyCode { get; set; }
+        public byte KeyCode { get; set; }
         [Key(1)]
         public uint EventCode { get; set; }
     }
